pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=7.2.1",
    "groq>=0.15.0",
    "openai>=1.59.9",
    "phidata>=2.7.9",
    "yfinance>=0.2.52",
]
